ticket:1849 - fix export of large data
Summary: there was an overflow issue when copying data to the result

Test Plan: systemtest

Reviewers: vmerticariu

Differential Revision: http://codereview.rasdaman.org/D704

+++ raslib/mddtypes.hh
@@ -53,7 +53,7 @@ rasdaman GmbH.
 //@Args: as unsigned int
 //@Memo: Module: {\bf raslib}.
 
-typedef unsigned int r_Bytes;
+typedef size_t r_Bytes;
 
 /**
   {\tt typedef unsigned int r_Bytes;}
